FIX: always include scrambler 0 / ZZZ in menu
otherwise drops will be offset from expected positions. I.e. if we created a menu where the window position before encoding was 'BFT', this means that the position of the identity scrambler in the Bombe needs to be set at BFT, AND that the identity scrambler means position 0. If the menu didn't include scrambler 0, and say started from scrambler 2, this already corresponds to window setting BFV, not BFT. Instead of calculating an offset for whatever scrambler happens to be lowest in the menu, seems more straightforward to just always include scrambler 0 / ZZZ